Secure actions on character

+++ Services/Character/deleteCharacter.php
@@ -12,23 +12,31 @@ session_start();
 if (isset($_POST["id"]) && isset($_SESSION["user"]))
 {
     $isSu = include_once "Services/Auth/isSu.php";
+    $charCreator = include_once dirname(__FILE__) . "/../Character/getCharacterCreatorId.php";
 
-    $character= new Character();
-    $character->id = mysqli_real_escape_string($connection, $_POST["id"]);
+    if ($charCreator == json_decode($_SESSION["user"])->id || $isSu) {
+        $character= new Character();
+        $character->id = mysqli_real_escape_string($connection, $_POST["id"]);
 
-    $query = sprintf("DELETE FROM %s WHERE characters.id = %d;",
-        Character::TABLE_NAME, $character->id);
+        $query = sprintf("DELETE FROM %s WHERE characters.id = %d;",
+            Character::TABLE_NAME, $character->id);
 
-    if (mysqli_query($connection, $query))
-    {
-        http_response_code(200);
+        if (mysqli_query($connection, $query))
+        {
+            http_response_code(200);
+        }
+        else
+        {
+            echo json_encode(array(
+                "error" => mysqli_error($connection),
+                "query" => $query));
+        }
     }
     else
     {
-        echo json_encode(array(
-            "error" => mysqli_error($connection),
-            "query" => $query));
+        echo json_encode(array("error" => "Must be creator or super user!"));
     }
+
 }
 else
 {
